Move definition of OCIAPIConfig and OCIAPI to types.
I needed to do this because I am trying to add a new 'OCIAPI' type
of 'noci' (nested oci) and wanted to hook it into api.NewOCIAPI.

Doing so would cause me an import loop, where
   * api.NewOCIApi woudl have to import noci to call noci.NewRepo
   * noci.NewRepo has to import and use pkg/api because
     it needs to return a api.OCIAPI
