+++ cmd/ocidist/cmd/repos.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 
 	"github.com/raharper/ocidist/pkg/api"
+	"github.com/raharper/ocidist/pkg/types"
 
 	"github.com/spf13/cobra"
 )
@@ -42,7 +43,7 @@ func doRepos(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	config := &api.OCIAPIConfig{TLSVerify: tlsVerify}
+	config := &types.OCIAPIConfig{TLSVerify: tlsVerify}
 	ociApi, err := api.NewOCIAPI(rawURL, config)
 	if err != nil {
 		return err
